requirements.txt
django==5.1.5
djangorestframework==3.15.2
django-filter==24.3